--- a/janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/MultiQueryPositions.java
+++ b/janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/MultiQueryPositions.java
@@ -1,0 +1,28 @@
+// Copyright 2023 JanusGraph Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.janusgraph.graphdb.tinkerpop.optimize;
+
+import org.apache.tinkerpop.gremlin.process.traversal.Step;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class MultiQueryPositions {
+
+    public List<Step> firstLoopMultiQueryStepLocations = new ArrayList<>();
+    public List<Step> currentLoopMultiQueryStepLocations = new ArrayList<>();
+    public Step nextLoopMultiQueryStepLocation;
+
+}