… docs/changelog.md
… docs/configs/janusgraph-cfg.md
… docs/operations/batch-processing.md
… janusgraph-backend-testutils/src/main/java/org/janusgraph/graphdb/JanusGraphTest.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/JanusGraphTraversalUtil.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/step/JanusGraphVertexStep.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/step/MultiQueriable.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/step/fetcher/BatchProcessingQueue.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/step/fetcher/MultiQueriableStepBatchFetcher.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/step/fetcher/PropertiesStepBatchFetcher.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/step/fetcher/VertexStepBatchFetcher.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/strategy/JanusGraphLocalQueryOptimizerStrategy.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/strategy/JanusGraphMultiQueryStrategy.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/strategy/JanusGraphUnusedMultiQueryRemovalStrategy.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/strategy/MultiQueriableStepRegistrationConsumer.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/optimize/strategy/MultiQueryStrategyRepeatStepMode.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/util/JanusGraphTraverserUtil.java
… janusgraph-core/src/test/java/org/janusgraph/graphdb/tinkerpop/optimize/step/fetcher/BatchProcessingQueueTest.java
… janusgraph-core/src/test/java/org/janusgraph/graphdb/tinkerpop/optimize/step/fetcher/MultiQueriableStepBatchFetcherTest.java
… janusgraph-core/src/test/java/org/janusgraph/graphdb/util/JanusGraphTraverserUtilTest.java
… janusgraph-core/src/test/java/org/janusgraph/graphdb/util/TestTraverserWithoutLoopsSupport.java
… janusgraph-inmemory/src/test/java/org/janusgraph/graphdb/inmemory/InMemoryGraphTest.java
… janusgraph-test/src/test/java/org/janusgraph/graphdb/tinkerpop/optimize/JanusGraphMultiQueryStepTest.java
… janusgraph-test/src/test/java/org/janusgraph/graphdb/tinkerpop/optimize/JanusGraphMultiQueryStrategyTest.java
… janusgraph-test/src/test/java/org/janusgraph/graphdb/tinkerpop/optimize/JanusGraphStepStrategyTest.java