This commit improves JanusGraphMultiQueryStrategy to better support multi-query compatible parent steps.

1. This commit brings better support for `repeat` step by introducing next itaration registration process.
Previously `repeat` children steps was getting traversers registered from the beginning of all outer repeat steps which
could result in duplicate or unnecesary retrievals for the first batch. Moreover, next iterations were not considered.
This commit changes the approach to bring different `repeat` step modes which can change the batches registration behaviour
to aacount only the closest `repeat` step, all `repeat` steps, all only starts of all `repeat` steps.

2. This commit adds support to almost all known TinkerPop Parent steps.
The exception is `match` step. We didn't have proper outter start registration for `match` step previously and now as well.

To not introduce a behavior change in multi-nested `repeat` steps the configuration option `query.batch.repeat-step-mode` is added with 3 modes available (`closest_repeat_parent` - mode which takes the closest `repeat` parent into account only, `starts_only_of_all_repeat_parents` - mode which takes all `repeat` parent steps but only starts into account (_previous behavior_), `all_repeat_parents` - same as `starts_only_of_all_repeat_parents` but takes `repeat` ends into account as well (_new default behavior_)).

The previous approach of inserting and finding multiQuery steps were split on 2 phases:
1. Insert multi-query step to appropriate location.
2. Find appropriate location.

Current approach is:
1. Insert multi-query step everywhere it might potentially be helpful.
2. Find appropriate location.

See performance improvement in this PR [here](https://github.com/JanusGraph/janusgraph/pull/3783#issuecomment-1565217154).

Fixes #3733
Fixes #3735
Fixes #2996

-----

Thank you for contributing to JanusGraph!

In order to streamline the review of the contribution we ask you
to ensure the following steps have been taken:

### For all changes:
- [x] Is there an issue associated with this PR? Is it referenced in the commit message?
- [x] Does your PR body contain #xyz where xyz is the issue number you are trying to resolve?
- [x] Has your PR been rebased against the latest commit within the target branch (typically `master`)?
- [x] Is your initial contribution a single, squashed commit?

### For code changes:
- [x] Have you written and/or updated unit tests to verify your changes?
- [ ] If adding new dependencies to the code, are these dependencies licensed in a way that is compatible for inclusion under [ASF 2.0](https://www.apache.org/legal/resolved.html#category-a)?
- [ ] If applicable, have you updated the LICENSE.txt file, including the main LICENSE.txt file in the root of this repository?
- [ ] If applicable, have you updated the NOTICE.txt file, including the main NOTICE.txt file found in the root of this repository?

### For documentation related changes:
- [x] Have you ensured that format looks appropriate for the output in which it is rendered?