@porunov 2023-05-27
Executed added benchmarks both for this PR and for `master` branch. The result are below.

**Master branch:**
```
Benchmark                                                      (fanoutFactor)  Mode  Cnt      Score     Error  Units
CQLMultiQueryBenchmark.getAllElementsTraversedFromOuterVertex             100  avgt    5     78.099 ±   5.815  ms/op
CQLMultiQueryBenchmark.getAllElementsTraversedFromOuterVertex             500  avgt    5   2027.809 ±  48.320  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingEmitRepeatSteps                100  avgt    5   1655.942 ±  54.545  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingEmitRepeatSteps                500  avgt    5  41937.395 ± 928.357  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingRepeatUntilSteps               100  avgt    5    819.768 ±  16.683  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingRepeatUntilSteps               500  avgt    5  20423.948 ± 269.510  ms/op
CQLMultiQueryBenchmark.getIdToOutVerticesProjection                       100  avgt    5     18.711 ±   2.841  ms/op
CQLMultiQueryBenchmark.getIdToOutVerticesProjection                       500  avgt    5    308.186 ±   7.184  ms/op
CQLMultiQueryBenchmark.getNames                                           100  avgt    5     96.689 ±   8.848  ms/op
CQLMultiQueryBenchmark.getNames                                           500  avgt    5   2540.583 ± 191.119  ms/op
CQLMultiQueryBenchmark.getNeighborNames                                   100  avgt    5     96.305 ±   6.908  ms/op
CQLMultiQueryBenchmark.getNeighborNames                                   500  avgt    5   2466.490 ±  91.254  ms/op
CQLMultiQueryBenchmark.getVerticesFilteredByAndStep                       100  avgt    5     25.953 ±   1.640  ms/op
CQLMultiQueryBenchmark.getVerticesFilteredByAndStep                       500  avgt    5    460.648 ±  27.039  ms/op
CQLMultiQueryBenchmark.getVerticesWithCoalesceUsage                       100  avgt    5     24.027 ±   2.691  ms/op
CQLMultiQueryBenchmark.getVerticesWithCoalesceUsage                       500  avgt    5    425.427 ±  36.108  ms/op
CQLMultiQueryBenchmark.getVerticesWithDoubleUnion                         100  avgt    5     10.331 ±   0.944  ms/op
CQLMultiQueryBenchmark.getVerticesWithDoubleUnion                         500  avgt    5    195.654 ±  24.556  ms/op
# Run complete. Total time: 00:22:51
```

**This PR:**
```
Benchmark                                                      (fanoutFactor)  Mode  Cnt     Score     Error  Units
CQLMultiQueryBenchmark.getAllElementsTraversedFromOuterVertex             100  avgt    5    79.839 ±   1.505  ms/op
CQLMultiQueryBenchmark.getAllElementsTraversedFromOuterVertex             500  avgt    5  2069.551 ±  11.601  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingEmitRepeatSteps                100  avgt    5   172.839 ±   6.408  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingEmitRepeatSteps                500  avgt    5  4471.290 ± 283.744  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingRepeatUntilSteps               100  avgt    5    91.972 ±   5.487  ms/op
CQLMultiQueryBenchmark.getElementsWithUsingRepeatUntilSteps               500  avgt    5  2384.490 ± 125.744  ms/op
CQLMultiQueryBenchmark.getIdToOutVerticesProjection                       100  avgt    5     5.442 ±   0.597  ms/op
CQLMultiQueryBenchmark.getIdToOutVerticesProjection                       500  avgt    5    93.286 ±   4.392  ms/op
CQLMultiQueryBenchmark.getNames                                           100  avgt    5   101.769 ±   8.236  ms/op
CQLMultiQueryBenchmark.getNames                                           500  avgt    5  2586.406 ±  54.791  ms/op
CQLMultiQueryBenchmark.getNeighborNames                                   100  avgt    5    98.194 ±   2.311  ms/op
CQLMultiQueryBenchmark.getNeighborNames                                   500  avgt    5  2477.925 ± 119.731  ms/op
CQLMultiQueryBenchmark.getVerticesFilteredByAndStep                       100  avgt    5    11.203 ±   1.394  ms/op
CQLMultiQueryBenchmark.getVerticesFilteredByAndStep                       500  avgt    5   235.142 ±  17.712  ms/op
CQLMultiQueryBenchmark.getVerticesWithCoalesceUsage                       100  avgt    5     9.140 ±   0.661  ms/op
CQLMultiQueryBenchmark.getVerticesWithCoalesceUsage                       500  avgt    5   182.895 ±  10.338  ms/op
CQLMultiQueryBenchmark.getVerticesWithDoubleUnion                         100  avgt    5    10.695 ±   1.994  ms/op
CQLMultiQueryBenchmark.getVerticesWithDoubleUnion                         500  avgt    5   198.177 ±  30.259  ms/op
# Run complete. Total time: 00:16:53
```

There were no performance difference for tests: `getAllElementsTraversedFromOuterVertex`, `getNames`, `getNeighborNames`, and `getVerticesWithDoubleUnion`.
There is good noticeable performance improvements for the next tests:
- `getElementsWithUsingEmitRepeatSteps` - `~9.5` times better performance in this PR.
- `getElementsWithUsingRepeatUntilSteps` - `~8.5` times better performance in this PR.
- `getIdToOutVerticesProjection` - `~3.4` times better performance in this PR.
- `getVerticesFilteredByAndStep` - `~2` times better performance in this PR.
- `getVerticesWithCoalesceUsage` - `~2.3` times better performance in this PR.

**Conclusion**: As this PR adds support to a number of new parent steps, it's expected that new parent steps will gain better performance in multi-query cases. As `repeat` step is now registering vertices with the start of `emit` and `until` steps it improved performance significantly for cases when `emit` or `until` use multiQueriable start steps.